// types.ts
export interface ZohoEmailListItem {
    messageId: string;
    from: { emailAddress: string; name: string; };
    to: { emailAddress: string; name: string; }[];
    subject: string;
    summary: string;
    receivedTime: string; // ISO String
    isRead: boolean;
}

export interface ZohoEmail extends ZohoEmailListItem {
    content: string; // HTML content
    attachments: {
        attachmentId: string;
        fileName: string;
        size: number;
    }[];
}